+++ rpc/src/server/mod.rs
@@ -14,8 +14,9 @@ use riker::actors::ActorSystem;
 use slog::Logger;
 
 use crypto::hash::{BlockHash, HashType};
-use storage::persistent::PersistentStorage;
 use shell::shell_channel::ShellChannelRef;
+use storage::persistent::PersistentStorage;
+use tezos_wrapper::TezosApiConnectionPool;
 
 use crate::empty;
 use crate::rpc_actor::{RpcCollectedStateRef, RpcServerRef};
@@ -43,11 +44,30 @@ pub struct RpcServiceEnvironment {
     shell_channel: ShellChannelRef,
     #[get = "pub(crate)"]
     log: Logger,
+    #[get = "pub(crate)"]
+    tezos_readonly_api: Arc<TezosApiConnectionPool>,
 }
 
 impl RpcServiceEnvironment {
-    pub fn new(sys: ActorSystem, actor: RpcServerRef, shell_channel: ShellChannelRef, persistent_storage: &PersistentStorage, genesis_hash: &BlockHash, state: RpcCollectedStateRef, log: &Logger) -> Self {
-        Self { sys, actor, shell_channel: shell_channel.clone(), persistent_storage: persistent_storage.clone(), genesis_hash: HashType::BlockHash.bytes_to_string(genesis_hash), state, log: log.clone() }
+    pub fn new(
+        sys: ActorSystem,
+        actor: RpcServerRef,
+        shell_channel: ShellChannelRef,
+        persistent_storage: &PersistentStorage,
+        tezos_readonly_api: Arc<TezosApiConnectionPool>,
+        genesis_hash: &BlockHash,
+        state: RpcCollectedStateRef,
+        log: &Logger) -> Self {
+        Self {
+            sys,
+            actor,
+            shell_channel: shell_channel.clone(),
+            persistent_storage: persistent_storage.clone(),
+            genesis_hash: HashType::BlockHash.bytes_to_string(genesis_hash),
+            state,
+            log: log.clone(),
+            tezos_readonly_api,
+        }
     }
 }
 
@@ -57,7 +77,7 @@ pub type Query = HashMap<String, Vec<String>>;
 
 pub type HResult = Result<Response<Body>, Box<dyn std::error::Error + Sync + Send>>;
 
-pub type Handler = Arc<dyn Fn(Request<Body>, Params, Query, RpcServiceEnvironment) -> Box<dyn Future<Output = HResult> + Send> + Send + Sync>;
+pub type Handler = Arc<dyn Fn(Request<Body>, Params, Query, RpcServiceEnvironment) -> Box<dyn Future<Output=HResult> + Send> + Send + Sync>;
 
 
 /// Spawn new HTTP server on given address interacting with specific actor system